now read from file if exists

+++ get_names_from_ask_fred.py
@@ -1,6 +1,7 @@
 import pandas as pd
 import requests
 from bs4 import BeautifulSoup
+import os
 
 base_url = 'https://askfred.net/Results/past.php?f%5Bevent_weapon_eq%5D=&f%5Bevent_gender_eq%5D=men&f%5Bevent_age_eq%5D=senior-open&f%5Bradius_mi%5D=300&vals%5Bloc%5D=&f%5Bname_contains%5D=&ops%5Bdate%5D=start_date_eq&vals%5Bdate%5D=&f%5Bevent_is_team%5D=&f%5Bevent_entries_gte%5D=&ops%5Bevent_rating%5D=event_rating_eq&vals%5Bevent_rating%5D=&f%5Bdivision_id%5D=&page_id={}'
 
@@ -21,23 +22,27 @@ def get_event_info(tournament_page):
 
 
 if __name__ == '__main__':
-    tournament_result_links = []
-    i = 1 # page count
-    while True:
-        response = requests.get(base_url.format(i))
-        soup = BeautifulSoup(response.text, 'html.parser')
-        
-        if soup.find('h3') is None:
-            # still tournaments being shown
-            tournament_result_links += [a['href'] for a in soup.find('table', {'id': 'past-tours'}).findAll('a') if 'results' in a['href']]
-        else:
-            # end of all tournaments in history for search, end
-            break
+    if os.path.exists('tournament_results_links.txt'):
+        with open('tournament_results_links.txt', 'r') as f:
+            tournament_result_links = f.read().split('\n')
+    else:
+        tournament_result_links = []
+        i = 1 # page count
+        while True:
+            response = requests.get(base_url.format(i))
+            soup = BeautifulSoup(response.text, 'html.parser')
+            
+            if soup.find('h3') is None:
+                # still tournaments being shown
+                tournament_result_links += [a['href'] for a in soup.find('table', {'id': 'past-tours'}).findAll('a') if 'results' in a['href']]
+            else:
+                # end of all tournaments in history for search, end
+                break
 
-        i += 1 # increment page count
+            i += 1 # increment page count
 
-    with open('touranment_result_links.txt', 'w') as f:
-        f.write('\n'.join(tournament_result_links))
+        with open('touranment_result_links.txt', 'w') as f:
+            f.write('\n'.join(tournament_result_links))
 
     dfs = []
     for link in tournament_result_links:
